Revert "Merge branch 'main' into NMS-Data-Model-classes"
This reverts commit 0b167d4d92f0b79635953313c6232347e33df9e5, reversing
changes made to f6a2137d06af8a4b886521428bad2a418afa318f.
