+++ DERMS/Common/AbstractModel/ModelDefines.cs
@@ -97,22 +97,22 @@ namespace Common.AbstractModel
         DER_ACTIVEPOWER             = 0x1121100000000205,
         DER_NOMINALPOWER            = 0x1121100000000305,
 
-        ENERGYSTORAGE                = 0x1121110000080000,
-        ENERGYSTORAGE_CAPACITY       = 0x1121110000080105,
-        ENERGYSTORAGE_STATEOFCHARGE  = 0x1121110000080205,
-        ENERGYSTORAGE_STATE          = 0x112111000008030A,
-        ENERGYSTORAGE_GENERATOR      = 0x1121110000080409,
+        DERENERGYSTORAGE                = 0x1121110000080000,
+        DERENERGYSTORAGE_CAPACITY       = 0x1121110000080105,
+        DERENERGYSTORAGE_STATEOFCHARGE  = 0x1121110000080205,
+        DERENERGYSTORAGE_STATE          = 0x112111000008030A,
+        DERENERGYSTORAGE_GENERATOR      = 0x1121110000000409,
 
-        GENERATOR                = 0x1121120000000000,
-        GENERATOR_DELTAPOWER     = 0x1121120000000105,
-        GENERATOR_ENERGYSTORAGE  = 0x1121120000000209,
+        DERGENERATOR                = 0x1121120000000000,
+        DERGENERATOR_DELTAPOWER     = 0x1121120000000105,
+        DERGENERATOR_STORAGE        = 0x1121120000000209,
 
-        SOLARGENERATOR           = 0x1121121000090000,
+        DERSOLARGENERATOR           = 0x1121121000090000,
 
-        WINDGENERATOR                = 0x11211210000A0000,
-        WINDGENERATOR_CUTOUTSPEED    = 0x11211210000A0105,
-        WINDGENERATOR_STARTUPSPEED     = 0x11211210000A0205,
-        WINDGENERATOR_NOMINALSPEED   = 0x11211210000A0305,
+        DERWINDGENERATOR                = 0x11211210000A0000,
+        DERWINDGENERATOR_CUTOUTSPEED    = 0x11211210000A0105,
+        DERWINDGENERATOR_CUTINSPEED     = 0x11211210000A0205,
+        DERWINDGENERATOR_NOMINALSPEED   = 0x11211210000A0305,
 
         EQUIPMENTCONTAINER              = 0x1111000000000000,
         EQUIPMENTCONTAINER_EQUIPEMENTS  = 0x1111000000000119,
