+++ DERMS/Importer/CIMAdapter/Importer/ImportHelper.cs
@@ -1,35 +1,14 @@
 ﻿using Common.AbstractModel;
-using Common.GDA;
 using System.Collections.Generic;
 
 namespace FTN.ESI.SIMES.CIM.CIMAdapter.Importer
 {
-    /// <summary>
-    /// Represents a <see cref="ResourceDescription"/> with a missing reference for given <see cref="ModelCode"/>, look at <see cref="ImportHelper.missingReferences"/>.
-    /// </summary>
-    internal struct MissingReferenceUnit
-    {
-        public ResourceDescription Rd;
-        public ModelCode ModelCode;
-
-        public MissingReferenceUnit(ResourceDescription rd, ModelCode modelCode)
-        {
-            this.Rd = rd;
-            this.ModelCode = modelCode;
-        }
-    }
-
 	/// <summary>
 	/// ImportHelper
 	/// </summary>
 	public class ImportHelper
 	{
 		private Dictionary<DMSType, int> typeCounter = new Dictionary<DMSType, int>();
-
-        /// <summary>
-        /// References such as one(optional) to one(mandatory)
-        /// </summary>
-        private Dictionary<string, MissingReferenceUnit> missingReferences = new Dictionary<string, MissingReferenceUnit>();
 		private Dictionary<string, long> rdfIDtoGIDMapping = new Dictionary<string, long>();
 
 
@@ -88,29 +67,5 @@ namespace FTN.ESI.SIMES.CIM.CIMAdapter.Importer
 			}
 			return gid;
 		}
-
-        /// <summary>
-        /// Remembers <see cref="ResourceDescription"/> which are missing references.
-        /// </summary>
-        /// <param name="rd">Which is missing a reference.</param>
-        /// <param name="modelCode">Of a missing reference.</param>
-        /// <param name="missingRdfId">RDFID of a missing entity.</param>
-        public void AddReferenceToMissingResourceDescription(ResourceDescription rd, ModelCode modelCode, string missingRdfId)
-        {
-            missingReferences.Add(missingRdfId, new MissingReferenceUnit(rd, modelCode));
-        }
-
-        /// <summary>
-        /// Adds missing references to resource descriptions added by <see cref="AddReferenceToMissingResourceDescription(ResourceDescription, ModelCode, string)"/>. 
-        /// </summary>
-        public void ExecuteReferenceAddition()
-        {
-            foreach (KeyValuePair<string, MissingReferenceUnit> missingReference in missingReferences)
-            {
-                MissingReferenceUnit unit = missingReference.Value;
-
-                unit.Rd.AddProperty(new Property(unit.ModelCode, rdfIDtoGIDMapping[missingReference.Key]));
-            }
-        }
 	}
 }
